+++ example/components/Braze.tsx
@@ -38,6 +38,14 @@ export const BrazeComponent = (): ReactElement => {
     useState<string>('bool');
   const [featureFlagPropertyKey, setFeatureFlagPropertyKey] = useState('');
 
+  // A comma-separated string of banner placement IDs.
+  const [requestedBannerPlacements, setRequestedBannerPlacements] = useState('placement_1, placement_2');
+
+  const [bannerPlacementId, setBannerPlacementId] = useState('');
+  const [bannerHeight, setBannerHeight] = useState(0);
+  const [displayedPlacement, setDisplayedPlacement] = useState('sdk-test-2');
+  const [displayedPlacementPlaceholder, setDisplayedPlacementPlaceholder] = useState('');
+
   const subscriptionStateButtons = useMemo(
     () => [
       {
@@ -210,6 +218,19 @@ export const BrazeComponent = (): ReactElement => {
       },
     );
 
+    const bannerCardsSubscription = Braze.addListener(
+      Braze.Events.BANNER_CARDS_UPDATED,
+      data => {
+        const banners = data.banners;
+        console.log(
+          `Received ${banners.length} Banner Cards with placement IDs:`,
+          banners.map(banner => banner.placementId),
+        );
+      },
+    );
+    // Perform an immediate refresh for default placements on app launch.
+    Braze.requestBannersRefresh(['placement_1', 'placement_2', 'sdk-test-2']);
+
     const featureFlagsSubscription = Braze.addListener(
       Braze.Events.FEATURE_FLAGS_UPDATED,
       flags => {
@@ -247,6 +268,7 @@ export const BrazeComponent = (): ReactElement => {
       listener.remove();
       inAppMessageSubscription.remove();
       contentCardsSubscription.remove();
+      bannerCardsSubscription.remove();
       featureFlagsSubscription.remove();
       sdkAuthErrorSubscription.remove();
       notificationListener.remove();
@@ -575,6 +597,40 @@ export const BrazeComponent = (): ReactElement => {
     }
   };
 
+  const getBannerByIdPress = async () => {
+    if (!bannerPlacementId) {
+      showToast('No Banner placement ID entered');
+      return;
+    }
+
+    const banner = await Braze.getBanner(bannerPlacementId);
+    if (!banner) {
+      showToast(`No Banner Card with placement ID: ${bannerPlacementId}`);
+      return;
+    }
+    showToast(`Found Banner Card. Check the console logs for details.`);
+    console.log(`Got Banner Card: ${JSON.stringify(banner, null, '\t')}`);
+  };
+
+  const requestBannersRefreshPress = () => {
+    if (!requestedBannerPlacements) {
+      console.log('No placement IDs entered for Banners refresh');
+      return;
+    }
+    // Split and trim the comma-separated string of placement IDs.
+    const bannerPlacements = requestedBannerPlacements.split(',').map(idString=>idString.trim());
+    Braze.requestBannersRefresh(bannerPlacements);
+    showToast('Banner Cards Refreshed');
+  };
+
+  const changeDisplayedBannerPress = async () => {
+    setDisplayedPlacement(displayedPlacementPlaceholder);
+    const banner = await Braze.getBanner(displayedPlacementPlaceholder);
+    if (!banner) {
+      showToast(`No banner found for placement: ${displayedPlacementPlaceholder}`);
+    }
+  };
+
   const requestPushPermission = () => {
     const options = {
       alert: true,
@@ -770,6 +826,62 @@ export const BrazeComponent = (): ReactElement => {
         <Text>Get Content Cards {'&'} Log interactions</Text>
       </TouchableHighlight>
 
+      {/* Banner Cards */}
+      <Space />
+      <View style={styles.row}>
+        <TextInput
+          style={styles.textInput}
+          placeholder='Placement IDs (comma-separated)'
+          autoCorrect={false}
+          autoCapitalize="none"
+          onChangeText={setRequestedBannerPlacements}
+          value={requestedBannerPlacements}
+        />
+        <TouchableHighlight onPress={requestBannersRefreshPress}>
+          <Text>Request Banners Refresh</Text>
+        </TouchableHighlight>
+      </View>
+      <View style={styles.row}>
+        <TextInput
+          style={styles.textInput}
+          placeholder='Placement ID'
+          autoCorrect={false}
+          autoCapitalize="none"
+          onChangeText={setBannerPlacementId}
+          value={bannerPlacementId}
+        />
+        <TouchableHighlight onPress={getBannerByIdPress}>
+          <Text>Get Banner Card by Placement ID</Text>
+        </TouchableHighlight>
+      </View>
+      <View style={styles.row}>
+        <TextInput
+          style={styles.textInput}
+          placeholder='Placement ID'
+          autoCorrect={false}
+          autoCapitalize="none"
+          onChangeText={(text) => setDisplayedPlacementPlaceholder(text)}
+          value={displayedPlacementPlaceholder}
+        />
+        <TouchableHighlight onPress={changeDisplayedBannerPress}>
+          <Text>Change Displayed Banner</Text>
+        </TouchableHighlight>
+      </View>
+      <Braze.BrazeBannerView
+        placementID={displayedPlacement}
+
+        // Uncomment the properties below for optional style overrides.
+        //
+        // In the simplest integration, only `placementID` is required.
+        // Dynamic height sizing is handled implicitly by the component without needing `onHeightChanged`.
+
+        // style={[styles.banner, { height: bannerHeight }]}
+        // onHeightChanged={height => {
+        //   console.log(`Banner height changed to: ${height}`);
+        //   setBannerHeight(height);
+        // }}
+      />
+
       {/* News Feed */}
 
       <Space />
@@ -872,6 +984,7 @@ const styles = StyleSheet.create({
     justifyContent: 'center',
     alignItems: 'center',
     backgroundColor: '#F5FCFF',
+    paddingTop: 100,
     paddingBottom: 100,
   },
   textInput: {