+++ plugin/ios/ExpoAdapterBraze/BrazeAppDelegate.swift
@@ -85,6 +85,10 @@ public class BrazeAppDelegate: ExpoAppDelegateSubscriber {
       if let useUUIDAsDeviceId = plistConfig["UseUUIDAsDeviceId"] as? Bool {
         configuration.useUUIDAsDeviceId = useUUIDAsDeviceId
       }
+        
+      if let forwardUniversalLinks = plistConfig["ForwardUniversalLinks"] as? Bool {
+        configuration.forwardUniversalLinks = forwardUniversalLinks
+      }
 
       let braze = BrazeReactBridge.perform(#selector(BrazeReactBridge.initBraze(_:)), with: configuration).takeUnretainedValue() as! Braze
       BrazeAppDelegate.braze = braze